pwnmcp/pyproject.toml
[project]
name = "pwnmcp-kiki1e"
version = "1.0.0"
description = "统一的PWN MCP服务器 - 整合静态分析、动态调试和漏洞利用功能"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "kiki1e"},
]

dependencies = [
    # MCP 核心
    "mcp>=1.0.0",
    
    # GDB 控制
    "pygdbmi>=0.11.0.0",
    
    # Web 服务
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    
    # PWN 工具
    "pwntools>=4.14.1",
    "ROPgadget>=6.5.0",
    "LibcSearcher",
    "ropper>=1.13.0",
    
    # 系统工具
    "psutil>=5.9.0",
    "rich>=14.1.0",
    
    # 其他依赖
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pwnmcp-server = "pwnmcp.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "mypy",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pwnmcp"]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "I"] # E: pycodestyle errors, F: pyflakes, I: isort
exclude = [
    ".venv",
    ".git",
    "__pycache__",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
exclude = ["venv", ".venv", "build", "dist"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
